@@ -1,7 +1,7 @@
 {
   "name": "express-handlebars",
   "description": "A Handlebars view engine for Express which doesn't suck.",
-  "version": "6.0.6",
+  "version": "6.0.7",
   "homepage": "https://github.com/express-handlebars/express-handlebars",
   "keywords": [
     "express",
@@ -24,7 +24,7 @@
     "node": ">=v12.22.9"
   },
   "dependencies": {
-    "glob": "^8.0.2",
+    "glob": "^8.1.0",
     "graceful-fs": "^4.2.10",
     "handlebars": "^4.7.7"
   },
@@ -33,25 +33,26 @@
     "example": "examples"
   },
   "devDependencies": {
-    "@semantic-release/changelog": "^6.0.1",
+    "@semantic-release/changelog": "^6.0.2",
     "@semantic-release/commit-analyzer": "^9.0.2",
     "@semantic-release/git": "^10.0.1",
-    "@semantic-release/github": "^8.0.4",
-    "@semantic-release/npm": "^9.0.1",
+    "@semantic-release/github": "^8.0.7",
+    "@semantic-release/npm": "^9.0.2",
     "@semantic-release/release-notes-generator": "^10.0.3",
-    "@types/glob": "^7.2.0",
-    "@types/jest": "^27.5.1",
-    "@typescript-eslint/eslint-plugin": "^5.23.0",
-    "@typescript-eslint/parser": "^5.23.0",
-    "eslint": "^8.15.0",
+    "@types/glob": "^8.0.1",
+    "@types/jest": "^29.4.0",
+    "@types/node": "^18.11.13",
+    "@typescript-eslint/eslint-plugin": "^5.49.0",
+    "@typescript-eslint/parser": "^5.49.0",
+    "eslint": "^8.32.0",
     "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.2.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "jest-cli": "^28.1.0",
-    "semantic-release": "^19.0.2",
-    "ts-jest": "^28.0.2",
-    "typescript": "^4.6.4"
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-n": "^15.6.1",
+    "eslint-plugin-promise": "^6.1.1",
+    "jest-cli": "^29.4.0",
+    "semantic-release": "^20.1.0",
+    "ts-jest": "^29.0.5",
+    "typescript": "^4.9.4"
   },
   "release": {
     "plugins": [
