@@ -1,6 +1,6 @@
 {
   "name": "socket.io-adapter",
-  "version": "2.4.0",
+  "version": "2.5.2",
   "license": "MIT",
   "repository": {
     "type": "git",
@@ -12,18 +12,23 @@
   "main": "./dist/index.js",
   "types": "./dist/index.d.ts",
   "description": "default socket.io in-memory adapter",
+  "dependencies": {
+    "ws": "~8.11.0"
+  },
   "devDependencies": {
+    "@types/mocha": "^10.0.1",
     "@types/node": "^14.11.2",
     "expect.js": "^0.3.1",
-    "mocha": "^8.1.3",
+    "mocha": "^10.2.0",
     "nyc": "^15.1.0",
-    "prettier": "^1.19.1",
-    "typescript": "^4.0.3"
+    "prettier": "^2.8.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^4.9.4"
   },
   "scripts": {
-    "test": "npm run format:check && tsc && nyc mocha test/index.js",
-    "format:check": "prettier --parser typescript --check 'lib/**/*.ts' 'test/**/*.js'",
-    "format:fix": "prettier --parser typescript --write 'lib/**/*.ts' 'test/**/*.js'",
+    "test": "npm run format:check && tsc && nyc mocha --require ts-node/register test/index.ts",
+    "format:check": "prettier --parser typescript --check 'lib/**/*.ts' 'test/**/*.ts'",
+    "format:fix": "prettier --parser typescript --write 'lib/**/*.ts' 'test/**/*.ts'",
     "prepack": "tsc"
   }
 }
