@@ -1,6 +1,6 @@
 {
   "name": "ws",
-  "version": "8.2.3",
+  "version": "8.11.0",
   "description": "Simple to use, blazing fast and thoroughly tested websocket client and server for Node.js",
   "keywords": [
     "HyBi",
@@ -17,8 +17,11 @@
   "license": "MIT",
   "main": "index.js",
   "exports": {
-    "import": "./wrapper.mjs",
-    "require": "./index.js"
+    ".": {
+      "import": "./wrapper.mjs",
+      "require": "./index.js"
+    },
+    "./package.json": "./package.json"
   },
   "browser": "browser.js",
   "engines": {
@@ -50,7 +53,7 @@
   "devDependencies": {
     "benchmark": "^2.1.4",
     "bufferutil": "^4.0.1",
-    "eslint": "^7.2.0",
+    "eslint": "^8.0.0",
     "eslint-config-prettier": "^8.1.0",
     "eslint-plugin-prettier": "^4.0.0",
     "mocha": "^8.4.0",
