@@ -1,15 +1,7 @@
-const multer = require('multer')
-const path = require('path')
+import {fileURLToPath} from 'url'
+import { dirname } from 'path'
 
-const storage = multer.diskStorage({
-    destination: (req, file, cb)=>{
-        cb(null, path.resolve(__dirname, '../public/img'))
-    },
-    filename: (req, file, cb)=>{
-        cb(null, file.originalname)
-    }
-})
+const __filename = fileURLToPath(import.meta.url)
+const __dirname = dirname(__filename)
 
-const uploader = multer({storage})
-
-module.exports = uploader
+export default __dirname;
